Updated how you add students and gradables, made a gradable type service that needs to be cleaned up a bit but does add, delete, and load gradable types for a loaded class. Next up is the gradable service
