+++ src/main/java/!run-compile.bat
@@ -1,0 +1,6 @@
+cd gui
+del /f *.class
+cd ..
+
+javac gui/*.java
+java gui/LogIn
