+++ src/main/java/gui/MainWindow.java
@@ -8,15 +8,12 @@ import java.awt.event.*;
 import java.awt.*;
 import javax.swing.table.DefaultTableModel;
 import java.util.ArrayList;
+import java.text.*;
 
 public class MainWindow {
 		
 		public MainWindow(final JFrame mainframe,final Data data) {
 			System.out.println("MainWindow to do list:");
-			System.out.println("> Add popup for addGradable");
-			System.out.println("> Add functionality to dropGradable");
-			System.out.println("> Add popup for addStudent");
-			System.out.println("> Add functionality to dropStudent");
 			System.out.println("> Add functionality to Save Class");
 			System.out.println("> Add functionality to Load Class");
 			System.out.println();
@@ -33,15 +30,23 @@ public class MainWindow {
 			JMenuItem menuItem_load = new JMenuItem("Load Class");
 			menu.add(menuItem_load);
 			
-			JMenuItem menuItem_Class_Profile = new JMenuItem("Class Profile");
-			menu.add(menuItem_Class_Profile);
-			
 			JMenuItem menuItem_exit = new JMenuItem("Exit");
 			menu.add(menuItem_exit);
 			
 
 			mainframe.setJMenuBar(menuBar);
 			// END Menu Toolbar
+			
+			// START Layout
+			final JPanel mainPanel = new JPanel();
+			mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
+			
+			JPanel treePanel = new JPanel();
+			treePanel.setLayout(new GridLayout(1,0));
+			JPanel buttonPanel = new JPanel();
+			buttonPanel.setLayout(new GridLayout(2,2));
+			
+			// END Layout
 			
 			// START Gradables tree
 			DefaultMutableTreeNode topGradables = new DefaultMutableTreeNode("Gradables");
@@ -67,20 +72,14 @@ public class MainWindow {
 
 			final JTree gradablesTree = new JTree(topGradables);
 			gradablesTree.getSelectionModel().setSelectionMode(TreeSelectionModel.SINGLE_TREE_SELECTION);
+			DefaultTreeModel gradablesModel = (DefaultTreeModel) gradablesTree.getModel();
+			DefaultMutableTreeNode gradablesRoot = (DefaultMutableTreeNode)gradablesModel.getRoot();
 			JScrollPane gradableView = new JScrollPane(gradablesTree);
 			
 			for(int i=0;i<gradablesTree.getRowCount();i++){
 				gradablesTree.expandRow(i);
 			}
 			
-			gradablesTree.addMouseListener(new MouseAdapter() {
-			public void mouseClicked(MouseEvent e) {
-				if (e.getClickCount() == 2) {
-					System.out.println((DefaultMutableTreeNode)gradablesTree.getLastSelectedPathComponent());
-				}
-			}
-		});
-		
 			// END Gradables tree 
 			
 			// START Students tree
@@ -102,13 +101,15 @@ public class MainWindow {
 				}
 			}
 			
-			
 			for (int i=0; i<studentCategories.size(); i++) {
 				topStudents.add(studentCategories.get(i));
 			}
 			
 			final JTree studentsTree = new JTree(topStudents);
 			studentsTree.getSelectionModel().setSelectionMode(TreeSelectionModel.SINGLE_TREE_SELECTION);
+			DefaultTreeModel studentsModel = (DefaultTreeModel) studentsTree.getModel();
+			DefaultMutableTreeNode studentsRoot = (DefaultMutableTreeNode)studentsModel.getRoot();
+				
 			JScrollPane studentView = new JScrollPane(studentsTree);
 			
 			for(int i=0;i<studentsTree.getRowCount();i++){
@@ -132,16 +133,11 @@ public class MainWindow {
 			// END Buttons
 			
 			// START layout
-			final JPanel mainPanel = new JPanel();
-			mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
 			
-			JPanel treePanel = new JPanel();
-			treePanel.setLayout(new GridLayout(0,2));
 			treePanel.add(gradableView);
 			treePanel.add(studentView);
 			
-			JPanel buttonPanel = new JPanel();
-			buttonPanel.setLayout(new GridLayout(2,2));
+			
 			buttonPanel.add(addGradable);
 			buttonPanel.add(addStudent);
 			buttonPanel.add(dropGradable);
@@ -171,7 +167,7 @@ public class MainWindow {
 					   ClassProfile c = new ClassProfile(mainframe, data);
 					   }
 					};
-			menuItem_Class_Profile.addActionListener(ClassProfileListener);
+			classSummary.addActionListener(ClassProfileListener);
 			
 			studentsTree.addMouseListener(new MouseAdapter() {
 				public void mouseClicked(MouseEvent e) {
@@ -205,87 +201,329 @@ public class MainWindow {
 			
 			ActionListener AddStudentListener = new ActionListener() {
 				public void actionPerformed(ActionEvent e) {
-					// System.out.println("Add Student");
-					NewStudentDialog nsd = new NewStudentDialog(data);
-					nsd.setModal(true);
-					nsd.showDialog();
-					ArrayList<Student> addedStudents = nsd.getStudents();
 					
-					DefaultTreeModel studentsModel = (DefaultTreeModel) studentsTree.getModel();
-					DefaultMutableTreeNode studentsRoot = (DefaultMutableTreeNode)studentsModel.getRoot();
+					treePanel.remove(1);
 					
-					Student addedStudent;
+					// START setup layout
+					JPanel newStudentPanel = new JPanel();		
+					newStudentPanel.setLayout(new BoxLayout(newStudentPanel, BoxLayout.Y_AXIS));
 					
-					for (int s=0; s<addedStudents.size(); s++) {
+					JPanel gridPanel = new JPanel();
+					gridPanel.setLayout(new GridLayout(4,2));
 					
-						addedStudent = addedStudents.get(s);
-						String sYear = addedStudent.getYear();
-						for (int i=0; i<studentsRoot.getChildCount();i++) {
-							if (studentsModel.getChild(studentsRoot,i).toString().equals(sYear)) {
-								DefaultMutableTreeNode yearBranch = (DefaultMutableTreeNode) studentsModel.getChild(studentsRoot,i);
-								studentsModel.insertNodeInto(new DefaultMutableTreeNode(addedStudent), yearBranch, yearBranch.getChildCount());
-							}
-						}
-					}
+					JPanel fnameInputPanel = new JPanel();
+					JPanel lnameInputPanel = new JPanel();
+					JPanel sidInputPanel = new JPanel();
+					JPanel yearInputPanel = new JPanel();
+					
+					JPanel addPanel = new JPanel();
+					JPanel botPanel = new JPanel();
+					// END setup layout
+					
+					Object[] yearOptions = data.studentTypes().toArray();
+					
+					final JTextField fnameTextField = new JTextField(10);
+					final JTextField lnameTextField = new JTextField(10);
+					final JTextField sidTextField = new JTextField(10);
+					final JComboBox<Object> yearCombo = new JComboBox<Object>(yearOptions);
+					
+					fnameInputPanel.add(fnameTextField);
+					lnameInputPanel.add(lnameTextField);
+					sidInputPanel.add(sidTextField);
+					yearInputPanel.add(yearCombo);
+					
+					gridPanel.add(new JLabel("First Name"));
+					gridPanel.add(fnameInputPanel);
+					gridPanel.add(new JLabel("Last Name"));
+					gridPanel.add(lnameInputPanel);
+					gridPanel.add(new JLabel("School ID"));
+					gridPanel.add(sidInputPanel);
+					gridPanel.add(new JLabel("Year"));
+					gridPanel.add(yearInputPanel);
+					
+					JButton addButton = new JButton("Add");
+					JButton addAndClose = new JButton("Add and Close");
+					JButton closeButton = new JButton("Close");
+					addPanel.add(addButton);
+					addPanel.add(addAndClose);
+					botPanel.add(addPanel);
+					botPanel.add(closeButton);
+					
+					newStudentPanel.add(gridPanel);
+					newStudentPanel.add(botPanel);
+					
+					treePanel.add(newStudentPanel,0,1);
+					mainframe.revalidate();
+					mainframe.repaint();
 					
 					for(int i=0;i<studentsTree.getRowCount();i++){
 						studentsTree.expandRow(i);
 					}	
+				
+					ActionListener addStudentButtonListener = new ActionListener(){
+						public void actionPerformed(ActionEvent e) {
+							String firstName = fnameTextField.getText();
+							String lastName = lnameTextField.getText();
+							String schoolID = sidTextField.getText();
+							String year = (String)yearCombo.getSelectedItem();
+							
+							// Check student ID against the database and error if there is a conflict
+							Student newStudent = new Student(firstName,lastName,schoolID,year);
+							for (int i=0; i<data.nGradables(); i++) {
+								newStudent.addGradable(data.getGradable(i));
+							}
+							
+							data.addStudent(newStudent);
+							
+							String sYear = newStudent.getYear();
+							for (int i=0; i<studentsRoot.getChildCount();i++) {
+								if (studentsModel.getChild(studentsRoot,i).toString().equals(sYear)) {
+									DefaultMutableTreeNode yearBranch = (DefaultMutableTreeNode) studentsModel.getChild(studentsRoot,i);
+									studentsModel.insertNodeInto(new DefaultMutableTreeNode(newStudent), yearBranch, yearBranch.getChildCount());
+								}
+							}
+							
+							fnameTextField.setText("");
+							lnameTextField.setText("");
+							sidTextField.setText("");
+						}
+					};
+					addButton.addActionListener(addStudentButtonListener);
+					
+					ActionListener addStudentAndCloseButtonListener = new ActionListener(){
+						public void actionPerformed(ActionEvent e) {
+							String firstName = fnameTextField.getText();
+							String lastName = lnameTextField.getText();
+							String schoolID = sidTextField.getText();
+							String year = (String)yearCombo.getSelectedItem();
+							
+							// Check student ID against the database and error if there is a conflict
+							Student newStudent = new Student(firstName,lastName,schoolID,year);
+							for (int i=0; i<data.nGradables(); i++) {
+								newStudent.addGradable(data.getGradable(i));
+							}
+							
+							data.addStudent(newStudent);
+												
+							String sYear = newStudent.getYear();
+							for (int i=0; i<studentsRoot.getChildCount();i++) {
+								if (studentsModel.getChild(studentsRoot,i).toString().equals(sYear)) {
+									DefaultMutableTreeNode yearBranch = (DefaultMutableTreeNode) studentsModel.getChild(studentsRoot,i);
+									studentsModel.insertNodeInto(new DefaultMutableTreeNode(newStudent), yearBranch, yearBranch.getChildCount());
+								}
+							}
+								
+							treePanel.remove(1);
+							treePanel.add(studentView,0,1);
+							mainframe.revalidate();
+							mainframe.repaint();
+						}
+					};
+					addAndClose.addActionListener(addStudentAndCloseButtonListener);
+					
+					ActionListener CloseAddStudentButtonListener = new ActionListener(){
+						public void actionPerformed(ActionEvent e) {
+							treePanel.remove(1);
+							treePanel.add(studentView,0,1);
+							mainframe.revalidate();
+							mainframe.repaint();
+							
+						}
+					};
+					closeButton.addActionListener(CloseAddStudentButtonListener);
+				
+				
+				
+				
 				}
 			};
 			addStudent.addActionListener(AddStudentListener);
 			
 			ActionListener AddGradableListener = new ActionListener() {
 				public void actionPerformed(ActionEvent e) {
-					System.out.println("Add Gradable");
 					
-					NewGradableDialog ngd = new NewGradableDialog(data);
-					ngd.setModal(true);
-					ngd.showDialog();
-					ArrayList<Gradable> addedGradables = ngd.getGradables();
+					treePanel.remove(0);
+					// START Setup Layout
+					JPanel newGradablePanel = new JPanel();
+					newGradablePanel.setLayout(new BoxLayout(newGradablePanel, BoxLayout.Y_AXIS));
 					
-					DefaultTreeModel gradablesModel = (DefaultTreeModel) gradablesTree.getModel();
-					DefaultMutableTreeNode gradablesRoot = (DefaultMutableTreeNode)gradablesModel.getRoot();
+					JPanel gridPanel = new JPanel();
+					gridPanel.setLayout(new GridLayout(4,2));
 					
-					Gradable addedGradable;
+					JPanel nameInputPanel = new JPanel();
+					JPanel pointsInputPanel = new JPanel();
+					JPanel weightInputPanel = new JPanel();
+					JPanel categoryInputPanel = new JPanel();
 					
-					for (int s=0; s<addedGradables.size(); s++) {
+					JPanel addPanel = new JPanel();
+					JPanel botPanel = new JPanel();
+					// END setup layout
+		
+					ArrayList<GradableType> gradableTypes = data.copyGradableTypes();
+					gradableTypes.add(new GradableType("New Category",0,0));
+					Object[] categoryOptions = gradableTypes.toArray();
 					
-						addedGradable = addedGradables.get(s);
-						for(int i=0; i<data.nStudents(); i++){
-							data.getStudent(i).addGradable(addedGradable);
-						}
-						
-						String gCategory = addedGradable.getType().getType();
-						boolean added = false;
-						for (int i=0; i<gradablesRoot.getChildCount();i++) {
-							if (gradablesModel.getChild(gradablesRoot,i).toString().equals(gCategory)) {
-								DefaultMutableTreeNode categoryBranch = (DefaultMutableTreeNode) gradablesModel.getChild(gradablesRoot,i);
-								gradablesModel.insertNodeInto(new DefaultMutableTreeNode(addedGradable), categoryBranch, categoryBranch.getChildCount());
-								added = true;
-							}
-						}
-						
-						if (!added){
-							int count = gradablesRoot.getChildCount();
-							gradablesModel.insertNodeInto(new DefaultMutableTreeNode(addedGradable.getType()),gradablesRoot,count);
-							DefaultMutableTreeNode categoryBranch = (DefaultMutableTreeNode) gradablesModel.getChild(gradablesRoot,count);
-							gradablesModel.insertNodeInto(new DefaultMutableTreeNode(addedGradable), categoryBranch, categoryBranch.getChildCount());
-						}
-					}
+					final JTextField nameTextField = new JTextField(10);
+					NumberFormat pointsFormat;
+					pointsFormat = NumberFormat.getNumberInstance();
+					final JFormattedTextField pointsTextField = new JFormattedTextField(pointsFormat);
+					pointsTextField.setValue(new Integer(100));
+					pointsTextField.setColumns(10);
+					// JTextField pointsTextField = new JTextField(10);
+					NumberFormat weightFormat;
+					weightFormat = NumberFormat.getNumberInstance();
+					final JFormattedTextField weightTextField = new JFormattedTextField(weightFormat);
+					weightTextField.setValue(new Integer(100));
+					weightTextField.setColumns(10);
+					weightTextField.setValue(new Integer(100));
+					// JTextField weightTextField = new JTextField(10);
+					
+					final JComboBox<Object> categoryCombo = new JComboBox<Object>(categoryOptions);
+					
+					nameInputPanel.add(nameTextField);
+					pointsInputPanel.add(pointsTextField);
+					weightInputPanel.add(weightTextField);
+					categoryInputPanel.add(categoryCombo);
+					
+					gridPanel.add(new JLabel("Gradable Name"));
+					gridPanel.add(nameInputPanel);
+					gridPanel.add(new JLabel("Total Points"));
+					gridPanel.add(pointsInputPanel);
+					gridPanel.add(new JLabel("Relative Weight"));
+					gridPanel.add(weightInputPanel);
+					gridPanel.add(new JLabel("Category"));
+					gridPanel.add(categoryInputPanel);
+					
+					JButton addButton = new JButton("Add");
+					JButton addAndClose = new JButton("Add and Close");
+					JButton closeButton = new JButton("Close");
+					addPanel.add(addButton);
+					addPanel.add(addAndClose);
+					botPanel.add(addPanel);
+					botPanel.add(closeButton);
+					
+					newGradablePanel.add(gridPanel);
+					newGradablePanel.add(botPanel);
+					
+					treePanel.add(newGradablePanel,0,0);
+					mainframe.revalidate();
+					mainframe.repaint();
+					
 					for(int i=0;i<gradablesTree.getRowCount();i++){
 						gradablesTree.expandRow(i);
 					}
+					
+					ActionListener addGradableButtonListener = new ActionListener(){
+						public void actionPerformed(ActionEvent e) {
+							String name = nameTextField.getText();
+							Integer points = ((Number)pointsTextField.getValue()).intValue();
+							Integer weight = ((Number)weightTextField.getValue()).intValue();
+							GradableType category = (GradableType)categoryCombo.getSelectedItem();
+							
+							Gradable newGradable = new Gradable(name,points,category,weight);
+							data.addGradable(newGradable);
+								
+							for(int i=0; i<data.nStudents(); i++){
+								data.getStudent(i).addGradable(newGradable);
+							}
+									
+								String gCategory = newGradable.getType().getType();
+								boolean added = false;
+								for (int i=0; i<gradablesRoot.getChildCount();i++) {
+									if (gradablesModel.getChild(gradablesRoot,i).toString().equals(gCategory)) {
+										DefaultMutableTreeNode categoryBranch = (DefaultMutableTreeNode) gradablesModel.getChild(gradablesRoot,i);
+										gradablesModel.insertNodeInto(new DefaultMutableTreeNode(newGradable), categoryBranch, categoryBranch.getChildCount());
+										added = true;
+									}
+								}
+								
+								if (!added){
+									int count = gradablesRoot.getChildCount();
+									gradablesModel.insertNodeInto(new DefaultMutableTreeNode(newGradable.getType()),gradablesRoot,count);
+									DefaultMutableTreeNode categoryBranch = (DefaultMutableTreeNode) gradablesModel.getChild(gradablesRoot,count);
+									gradablesModel.insertNodeInto(new DefaultMutableTreeNode(newGradable), categoryBranch, categoryBranch.getChildCount());
+								}
+
+							nameTextField.setText("");
+							pointsTextField.setText("");
+							weightTextField.setText("");
+						}
+					};
+					addButton.addActionListener(addGradableButtonListener);
+		
+					ActionListener addAndGradableCloseButtonListener = new ActionListener(){
+						public void actionPerformed(ActionEvent e) {
+							String name = nameTextField.getText();
+							Integer points = ((Number)pointsTextField.getValue()).intValue();
+							Integer weight = ((Number)weightTextField.getValue()).intValue();
+							GradableType category = (GradableType)categoryCombo.getSelectedItem();
+							
+							Gradable newGradable = new Gradable(name,points,category,weight);
+							data.addGradable(newGradable);
+								
+							for(int i=0; i<data.nStudents(); i++){
+								data.getStudent(i).addGradable(newGradable);
+							}
+									
+							String gCategory = newGradable.getType().getType();
+							boolean added = false;
+							for (int i=0; i<gradablesRoot.getChildCount();i++) {
+								if (gradablesModel.getChild(gradablesRoot,i).toString().equals(gCategory)) {
+									DefaultMutableTreeNode categoryBranch = (DefaultMutableTreeNode) gradablesModel.getChild(gradablesRoot,i);
+									gradablesModel.insertNodeInto(new DefaultMutableTreeNode(newGradable), categoryBranch, categoryBranch.getChildCount());
+									added = true;
+								}
+							}
+							
+							if (!added){
+								int count = gradablesRoot.getChildCount();
+								gradablesModel.insertNodeInto(new DefaultMutableTreeNode(newGradable.getType()),gradablesRoot,count);
+								DefaultMutableTreeNode categoryBranch = (DefaultMutableTreeNode) gradablesModel.getChild(gradablesRoot,count);
+								gradablesModel.insertNodeInto(new DefaultMutableTreeNode(newGradable), categoryBranch, categoryBranch.getChildCount());
+							}
+									
+							treePanel.remove(0);
+							treePanel.add(gradableView,0,0);
+							mainframe.revalidate();
+							mainframe.repaint();
+									
+						}
+					};
+					addAndClose.addActionListener(addAndGradableCloseButtonListener);
+					
+					ActionListener CloseAddGradableButtonListener = new ActionListener(){
+						public void actionPerformed(ActionEvent e) {
+							treePanel.remove(0);
+							treePanel.add(gradableView,0,0);
+							mainframe.revalidate();
+							mainframe.repaint();
+						}
+					};
+					closeButton.addActionListener(CloseAddGradableButtonListener);
+		
+					ActionListener categoryListener = new ActionListener(){
+							public void actionPerformed(ActionEvent e){
+								GradableType mySelection = (GradableType)categoryCombo.getSelectedItem();
+								if (mySelection.getType().equals("New Category")){
+									NewCategoryDialog ncd = new NewCategoryDialog(data);
+									ncd.setModal(true);
+									ncd.showDialog();
+									ArrayList<GradableType> addedGradableTypes = ncd.getGradableTypes();
+									for (int i=0;i<addedGradableTypes.size(); i++) {
+									categoryCombo.addItem(addedGradableTypes.get(i));
+									categoryCombo.setSelectedItem(addedGradableTypes.get(i));
+									}
+								}
+								}
+							};
+					categoryCombo.addActionListener( categoryListener );
 				}
 			};
 			addGradable.addActionListener(AddGradableListener);
 			
 			ActionListener DropStudentListener = new ActionListener() {
 				public void actionPerformed(ActionEvent e) {
-				// System.out.println("Drop Student");
-				DefaultMutableTreeNode node = (DefaultMutableTreeNode)studentsTree.getLastSelectedPathComponent();
-				// System.out.println((Student)node.getUserObject());
-				
+				DefaultMutableTreeNode node = (DefaultMutableTreeNode)studentsTree.getLastSelectedPathComponent();				
 				if (node.getUserObject() instanceof Student) {
 					data.dropStudent((Student)node.getUserObject());
 					DefaultTreeModel studentsModel = (DefaultTreeModel) studentsTree.getModel();
@@ -301,7 +539,6 @@ public class MainWindow {
 				
 				if (node.getUserObject() instanceof Gradable) {
 					data.dropGradable((Gradable)node.getUserObject());
-					// NEED TO DROP IT FROM EVERY STUDENT TOO
 					for (int i = 0; i < data.nStudents(); i++) {
 						data.getStudent(i).dropGradable((Gradable)node.getUserObject());
 					}
@@ -313,13 +550,8 @@ public class MainWindow {
 			};
 			dropGradable.addActionListener(DropGradableListener);
 			
-			classSummary.addActionListener(ClassProfileListener);
-			// END Action Listeners 
 			
+			// END Action Listeners 
 		}
-
-		
-	// private void drawMainWindow(JFrame mainframe,Data data) {
-
 	
 }
